+++ changelog.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## UNRELEASED
+- ✨ Added base config without prettier.
 
 ## v3.0.1 - 2024-09-29
 - 🐛 Added missing `eslint-formatter-checkstyle` package.